+++ playground/default-export/src/index.js
@@ -1,0 +1,7 @@
+function main() {
+  return true
+}
+
+export default main
+
+export const version = '1.0.0'