+++ src/node/core/pkg/loadPkgWithReporting.ts
@@ -1,6 +1,7 @@
 import chalk from 'chalk'
 import {ZodError} from 'zod'
 import {Logger} from '../../logger'
+import {assertFirst, assertLast, assertOrder} from './helpers'
 import {loadPkg} from './loadPkg'
 import {PackageJSON} from './types'
 
@@ -14,8 +15,43 @@ export async function loadPkgWithReporting(options: {
   try {
     const pkg = await loadPkg({cwd})
 
-    if (pkg.type === undefined) {
-      logger.warn('no "type" field in package.json, defaulting to "commonjs"')
+    // validate exports
+    if (pkg.exports) {
+      const _exports = Object.entries(pkg.exports)
+
+      for (const [expPath, exp] of _exports) {
+        if (typeof exp === 'string') {
+          continue
+        }
+
+        const keys = Object.keys(exp)
+
+        if (!assertFirst('types', keys)) {
+          logger.warn(`exports["${expPath}"]: the \`types\` property should be the first property`)
+        }
+
+        if (!assertOrder('require', 'import', keys)) {
+          logger.warn(
+            `exports["${expPath}"]: the \`require\` property should come before the \`import\` property`
+          )
+        }
+
+        if (!assertOrder('require', 'node', keys)) {
+          logger.warn(
+            `exports["${expPath}"]: the \`require\` property should come before the \`node\` property`
+          )
+        }
+
+        if (!assertOrder('node', 'import', keys)) {
+          logger.warn(
+            `exports["${expPath}"]: the \`node\` property should come before \`import\` property`
+          )
+        }
+
+        if (!assertLast('default', keys)) {
+          logger.warn(`exports["${expPath}"]: the \`default\` property should be the last property`)
+        }
+      }
     }
 
     return pkg